progress, but not done, loading xml data into db
some strange things in that xml stuff

questionable whether it is any good, which is why I am trying to break
it up

at the stage where I am trying to stuff the actual observations.
Testing is difficult with temp tables because I have foreign key
constraints.  have to set up another set of temp tables for the
"perl/xml" data so that the FK constraints can be removed (during test)
Then push them in and double check that everything is good, then
refactor the absolute mess I am making of parser.js